migrated device tools settings to sub-fragment
