+++ app/src/main/java/hpsaturn/pollutionreporter/view/SettingsDeviceTools.java
@@ -2,20 +2,45 @@ package hpsaturn.pollutionreporter.view;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.os.Handler;
 
 import androidx.annotation.Nullable;
+import androidx.preference.SwitchPreference;
+
+import com.hpsaturn.tools.Logger;
+
+import java.text.DecimalFormat;
 
 import hpsaturn.pollutionreporter.R;
+import hpsaturn.pollutionreporter.models.CommandConfig;
+import hpsaturn.pollutionreporter.models.GeoConfig;
 import hpsaturn.pollutionreporter.models.ResponseConfig;
+import io.nlopez.smartlocation.SmartLocation;
 
 /**
  * Created by Antonio Vanegas @hpsaturn on 3/31/21.
  */
 public class SettingsDeviceTools extends SettingsBaseFragment{
 
+    public static final String TAG = SettingsDeviceTools.class.getSimpleName();
+
 
     @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
+
+        if (!onSensorReading) {
+
+            if (key.equals(getString(R.string.key_setting_enable_reboot))) {
+                performRebootDevice();
+            } else if (key.equals(getString(R.string.key_setting_enable_clear))) {
+                performClearDevice();
+            } else if (key.equals(getString(R.string.key_setting_enable_location))) {
+                saveLocation();
+            }
+
+        }
+        else
+            Logger.i(TAG,"skyp onSharedPreferenceChanged because is in reading mode!");
 
     }
 
@@ -27,13 +52,104 @@ public class SettingsDeviceTools extends SettingsBaseFragment{
 
     @Override
     protected void refreshUI() {
-
+        Logger.i(TAG,"[Settings Tools] refreshUI");
+        updateLocationSummary();
+        lastLocation = SmartLocation.with(getActivity()).location().getLastLocation();
+        validateLocationSwitch();
     }
 
     @Override
     protected void onConfigRead(ResponseConfig config) {
-
+        updateLocationSummary();
+        validateLocationSwitch();
     }
+
+
+    private void validateLocationSwitch() {
+        SwitchPreference locationSwitch = findPreference(getString(R.string.key_setting_enable_location));
+        locationSwitch.setEnabled(lastLocation!=null);
+    }
+
+    private void saveLocation() {
+        SwitchPreference locationSwitch = findPreference(getString(R.string.key_setting_enable_location));
+        if(lastLocation != null) {
+            if(locationSwitch.isChecked()) {
+                snackBar = getMain().getSnackBar(R.string.msg_set_current_location, R.string.bt_location_save_action, view -> {
+                    getMain().showSnackMessage(R.string.msg_save_location);
+                    GeoConfig config = new GeoConfig();
+                    config.lat = lastLocation.getLatitude();
+                    config.lon = lastLocation.getLongitude();
+                    config.alt = lastLocation.getAltitude();
+                    config.spd = lastLocation.getSpeed();
+                    sendSensorConfig(config);
+                    Handler handler = new Handler();
+                    handler.postDelayed(() -> locationSwitch.setChecked(false), 2000);
+                });
+                snackBar.show();
+            }
+            else{
+                snackBar.dismiss();
+            }
+        }
+        else {
+            getMain().showSnackMessage(R.string.msg_save_location_failed);
+        }
+        updateLocationSummary();
+    }
+
+    private void updateLocationSummary() {
+        if (lastLocation != null) {
+            DecimalFormat precision = new DecimalFormat("0.000");
+            String accu = "Accu:" + (int) lastLocation.getAccuracy() + "m ";
+            String lat = "(" + precision.format(lastLocation.getLatitude());
+            String lon = "," + precision.format(lastLocation.getLongitude()) + ")";
+            updateSummary(R.string.key_setting_enable_location,accu + lat + lon);
+        }
+    }
+
+    private void performRebootDevice() {
+        SwitchPreference rebootSwitch = findPreference(getString(R.string.key_setting_enable_reboot));
+        if (!rebootSwitch.isChecked()) {
+            if(snackBar!=null)snackBar.dismiss();
+        } else {
+            snackBar = getMain().getSnackBar(R.string.bt_device_reboot, R.string.bt_device_reboot_action, view -> {
+                rebootSwitch.setChecked(false);
+                sendSensorReboot();
+                getMain().showSnackMessageSlow(R.string.msg_device_reboot);
+                Handler handler = new Handler();
+                handler.postDelayed(() -> getMain().finish(), 3000);
+            });
+            snackBar.show();
+        }
+    }
+
+    private void sendSensorReboot() {
+        CommandConfig config = new CommandConfig();
+        config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
+        config.act = "rbt";
+        sendSensorConfig(config);
+    }
+
+    private void performClearDevice() {
+        SwitchPreference clearSwitch = findPreference(getString(R.string.key_setting_enable_clear));
+        if (!clearSwitch.isChecked()) {
+            if(snackBar!=null)snackBar.dismiss();
+        } else {
+            snackBar = getMain().getSnackBar(R.string.bt_device_clear, R.string.bt_device_clear_action, view -> {
+                clearSwitch.setChecked(false);
+                CommandConfig config = new CommandConfig();
+                config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
+                config.act = "cls";
+                getMain().showSnackMessageSlow(R.string.msg_device_clear);
+                sendSensorConfig(config);
+                clearSharedPreferences();
+                Handler handler = new Handler();
+                handler.postDelayed(() -> getMain().finish(), 3000);
+            });
+            snackBar.show();
+        }
+    }
+
 
 
 }
